--- a/changelog.d/13575.misc
+++ b/changelog.d/13575.misc
@@ -1,0 +1,1 @@
+Optimize how Synapse calculates domains to fetch from during backfill.