--- a/synapse/handlers/room.py
+++ b/synapse/handlers/room.py
@@ -1465,7 +1465,7 @@ class TimestampLookupHandler:
             # Loop through each homeserver candidate until we get a succesful response
             for domain in likely_domains:
                 # We don't want to ask our own server for information we don't have
-                if domain != self.server_name:
+                if domain == self.server_name:
                     continue
 
                 try: