@DMRobertson 2022-09-29
```
$ git log --oneline -S get_users_in_room v1.66.0..v1.68.0
  e124b24ad 1.68.0rc1
  906cead9c Update docstrings to explain the impact of partial state (#13750)
  cdf7fb737 1.67.0rc1
  1a209efdb Update `get_users_in_room` mis-use to get hosts with dedicated `get_current_hosts_in_room` (#13605)
  d58615c82 Directly lookup local membership instead of getting all members in a room first (`get_users_in_room` mis-use) (#13608)
  c807b814a Use dedicated `get_local_users_in_room` to find local users when calculating `join_authorised_via_users_server` of a `/make_jo  in` request (#13606)
```

On the affected hosts we saw that postgres used a heap scan rather than an index scan for the aggressive query. Forcing an index-only scan (SET enable_bitmapscan to off) made the query quicker (500->300ms) for one trial set of parameters. Unclear if those are representitive. Also unclear if the index-only scan used more or less file IO.

But this might be barking up the wrong tree. Are we perhaps calling `get_users_in_room` in more circumstances now? Do we need to up its cache?